@@ -1,19 +1,27 @@
 {
   "name": "nojimage/cakephp-remember-me",
-  "description": "RememberMe authentication adapter plugin for CakePHP 3",
+  "description": "RememberMe authentication adapter plugin for CakePHP",
+  "keywords": [
+    "auth",
+    "authentication",
+    "cakephp",
+    "cookie",
+    "remember-me"
+  ],
   "type": "cakephp-plugin",
   "license": ["MIT"],
   "require": {
-    "php": ">=5.6",
+    "php": ">=7.2",
     "ext-json": "*",
-    "cakephp/core": "^3.5",
-    "cakephp/orm": "^3.5"
+    "cakephp/core": "^4.0",
+    "cakephp/orm": "^4.0",
+    "cakephp/authentication": "^2.0"
   },
   "require-dev": {
-    "cakephp/authentication": "^1.0",
-    "cakephp/cakephp": "^3.5",
-    "cakephp/cakephp-codesniffer": "^3.3",
-    "phpunit/phpunit": "^5.7|^6.0"
+    "cakephp/cakephp": "^4.0",
+    "cakephp/cakephp-codesniffer": "^4.6",
+    "cakephp/migrations": "^3.0",
+    "phpunit/phpunit": "^8.5 || ^9.3"
   },
   "autoload": {
     "psr-4": {
@@ -23,14 +31,19 @@
   "autoload-dev": {
     "psr-4": {
       "RememberMe\\Test\\": "tests",
-      "Cake\\Test\\": "./vendor/cakephp/cakephp/tests"
+      "TestApp\\": "tests/test_app/TestApp",
+      "Cake\\Test\\": "vendor/cakephp/cakephp/tests"
     }
   },
   "scripts": {
     "cs-check": "phpcs -p --extensions=php --standard=vendor/cakephp/cakephp-codesniffer/CakePHP ./src ./tests",
-    "cs-fix": "phpcbf -p --extensions=php --standard=vendor/cakephp/cakephp-codesniffer/CakePHP ./src ./tests"
+    "cs-fix": "phpcbf -p --extensions=php --standard=vendor/cakephp/cakephp-codesniffer/CakePHP ./src ./tests",
+    "update-lowest": "composer update --prefer-lowest --prefer-stable"
   },
   "config": {
-    "sort-packages": true
+    "sort-packages": true,
+    "allow-plugins": {
+      "dealerdirect/phpcodesniffer-composer-installer": true
+    }
   }
 }
