@@ -12,7 +12,7 @@ class CreateRememberMeTokens extends AbstractMigration
      * http://docs.phinx.org/en/latest/migrations.html#the-change-method
      * @return void
      */
-    public function change()
+    public function change(): void
     {
         $table = $this->table('remember_me_tokens');
         $table->addTimestamps('created', 'modified');
